requirements.txt
streamlit
streamlit-chat
websocket-client
